Finish functionality
The game stops once someone reaches 5 points correctly. I also
replaced the alerts with the outcome of each round with text injection
into the DOM. (don't even know if that's the term)

Once the game finishes, a text appears on the page telling the user that
it's possible to restart by refreshing the page.

I wanted to add a reset button that would reset the scores to 0
but I opted out for the above because of ease of implementation. Might
come back to this later and implement the reset button.

Right now I want to focus on learning new features so I didn't spend
time on a CSS on purpose.